feat: add intro  section and pre-commit configuration (#1)
* chore: add pre-commit configuration

Add pre-commit configuration file to enforce code quality checks
Fix any issues found by pre-commit hooks in existing files

* feat: add introduction section

Add an introduction section to the README file to provide context and purpose of the project.

* chore: add actionlint to pre-commit config

actionlint is a linter for GitHub Actions workflows.
It helps to ensure that the workflows are correctly configured and follow best practices.

* ci: Add pre-commit workflow

pre-commit workflow ensure code quality.
It runs on every push and pull request to the `main` branch.
Ensures that code adheres to the defined standards and checks.
Non-compliant code will not be merged.